Show calculated grades as matching if they match floored grades as well (fixes #102)
